[Era/SoD] Convert to JS Modules
This allows much better scoping and separation.

It will pave the way for importing era/class.js into sod/class.js so that we can only specify the changed bits. Same for era/naxx.js -> sod/naxx.js and so on.
