+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "js/ts.implicitProjectConfig.checkJs": true
+}
