+++ era/index.html
@@ -24,12 +24,8 @@
       async
       src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js"
     ></script>
-    <script src="colors.js"></script>
-    <script src="base.js"></script>
-    <script src="spells.js"></script>
     <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-    <script src="threat.js"></script>
-    <script src="front.js"></script>
+    <script src="main.js" type="module"></script>
 
     <meta name="viewport" content="width=device-width, initial-scale=1" />
 
@@ -107,7 +103,7 @@
       }
     </style>
   </head>
-  <body onload="loadPage()">
+  <body>
     <div id="input">
       <table>
         <tr>
