+++ sod/index.html
@@ -24,21 +24,8 @@
       async
       src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js"
     ></script>
-    <script src="../era/colors.js"></script>
-    <script src="../era/base.js"></script>
-    <script src="warrior.js"></script>
-    <script src="paladin.js"></script>
-    <script src="druid.js"></script>
-    <script src="rogue.js"></script>
-    <script src="hunter.js"></script>
-    <script src="warlock.js"></script>
-    <script src="shaman.js"></script>
-    <script src="mage.js"></script>
-    <script src="priest.js"></script>
-    <script src="spells.js"></script>
     <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-    <script src="../era/threat.js"></script>
-    <script src="../era/front.js"></script>
+    <script src="main.js" type="module"></script>
 
     <meta name="viewport" content="width=device-width, initial-scale=1" />
 
@@ -116,7 +103,7 @@
       }
     </style>
   </head>
-  <body onload="loadPage()">
+  <body>
     <div id="input">
       <table>
         <tr>
